pyproject.toml
[project]
name = "meetingroom-bot"
version = "0.1.0"
description = "Slack bot for meeting room reservations"
requires-python = ">=3.10"
dependencies = [
    "slack-bolt>=1.18.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "black>=23.0.0",
]
